Fjernet bruken av mangler og mangler navn etc i backend, dette bør være et frontend-styrt greie. Men har prøvd å kun gjøre det for de nye APIene, ikke for de gamle.
