+++ src/main/kotlin/no/nav/klage/oppgave/api/mapper/KlagebehandlingMapper.kt
@@ -4,22 +4,62 @@ package no.nav.klage.oppgave.api.mapper
 import no.nav.klage.oppgave.api.view.KlagebehandlingListView
 import no.nav.klage.oppgave.api.view.KlagebehandlingView
 import no.nav.klage.oppgave.api.view.KvalitetsvurderingView
+import no.nav.klage.oppgave.clients.egenansatt.EgenAnsattService
+import no.nav.klage.oppgave.clients.pdl.PdlFacade
 import no.nav.klage.oppgave.domain.elasticsearch.EsKlagebehandling
 import no.nav.klage.oppgave.domain.klage.Hjemmel
 import no.nav.klage.oppgave.domain.klage.Klagebehandling
+import no.nav.klage.oppgave.domain.klage.Mottak
 import no.nav.klage.oppgave.util.getLogger
 import no.nav.klage.oppgave.util.getSecureLogger
 import org.springframework.stereotype.Service
 
 @Service
 class KlagebehandlingMapper(
-
+    private val pdlFacade: PdlFacade,
+    private val egenAnsattService: EgenAnsattService
 ) {
 
     companion object {
         @Suppress("JAVA_CLASS_ON_COMPANION")
         private val logger = getLogger(javaClass.enclosingClass)
         private val secureLogger = getSecureLogger()
+    }
+
+    fun mapKlagebehandlingOgMottakToEsKlagebehandling(klagebehandlingOgMottak: Pair<Klagebehandling, Mottak>): EsKlagebehandling {
+
+        val klagebehandling = klagebehandlingOgMottak.first
+        //TODO: Nå bruker jeg ikke mottak her, så jeg kunne endret Pair<Klagebehandling, Mottak> til å bare være Klagebehandling?
+        //TODO: Er det noe vi skal indeksere opp som kommer fra Mottak? Beskrivelse f.eks?
+
+        val personInfo = klagebehandling.foedselsnummer?.let { pdlFacade.getPersonInfo(it) }
+        val erFortrolig = personInfo?.harBeskyttelsesbehovFortrolig() ?: false
+        val erStrengtFortrolig = personInfo?.harBeskyttelsesbehovStrengtFortrolig() ?: false
+        val erEgenAnsatt = klagebehandling.foedselsnummer?.let { egenAnsattService.erEgenAnsatt(it) } ?: false
+        val navn = personInfo?.navn
+
+        return EsKlagebehandling(
+            id = klagebehandling.id.toString(),
+            versjon = klagebehandling.versjon,
+            journalpostId = klagebehandling.saksdokumenter.map { it.referanse },
+            saksreferanse = klagebehandling.referanseId,
+            tildeltEnhet = klagebehandling.tildeltEnhet,
+            tema = klagebehandling.tema,
+            sakstype = klagebehandling.sakstype,
+            tildeltSaksbehandlerident = klagebehandling.tildeltSaksbehandlerident,
+            innsendt = klagebehandling.innsendt,
+            mottattFoersteinstans = klagebehandling.mottattFoersteinstans,
+            mottattKlageinstans = klagebehandling.mottattKlageinstans,
+            frist = klagebehandling.frist,
+            startet = klagebehandling.startet,
+            avsluttet = klagebehandling.avsluttet,
+            hjemler = klagebehandling.hjemler.map { it.original },
+            foedselsnummer = klagebehandling.foedselsnummer,
+            navn = navn,
+            egenAnsatt = erEgenAnsatt,
+            fortrolig = erFortrolig,
+            strengtFortrolig = erStrengtFortrolig
+        )
     }
 
     fun mapEsKlagebehandlingerToListView(
@@ -59,7 +99,7 @@ class KlagebehandlingMapper(
             fraNAVEnhet = klagebehandling.avsenderEnhetFoersteinstans,
             fraSaksbehandlerident = klagebehandling.avsenderSaksbehandleridentFoersteinstans,
             mottattFoersteinstans = klagebehandling.mottattFoersteinstans,
-            foedselsnummer = klagebehandling.foedselsnummer ?: "mangler",
+            foedselsnummer = klagebehandling.foedselsnummer,
             tema = klagebehandling.tema.id,
             sakstype = klagebehandling.sakstype.navn,
             mottatt = klagebehandling.mottattKlageinstans,
