+++ src/main/kotlin/no/nav/klage/oppgave/api/mapper/OppgaveMapper.kt
@@ -5,8 +5,6 @@ import no.nav.klage.oppgave.api.internal.OppgaveKopiAPIModel
 import no.nav.klage.oppgave.clients.egenansatt.EgenAnsattService
 import no.nav.klage.oppgave.clients.pdl.PdlFacade
 import no.nav.klage.oppgave.domain.elasticsearch.EsKlagebehandling
-import no.nav.klage.oppgave.domain.klage.Klagebehandling
-import no.nav.klage.oppgave.domain.klage.Mottak
 import no.nav.klage.oppgave.domain.oppgavekopi.*
 import no.nav.klage.oppgave.util.getLogger
 import no.nav.klage.oppgave.util.getSecureLogger
@@ -88,42 +86,6 @@ class OppgaveMapper(
             metadata = oppgave.metadata?.map { (k, v) ->
                 MetadataNoekkel.valueOf(k.name) to v
             }?.toMap() ?: emptyMap()
-        )
-    }
-
-    fun mapKlagebehandlingOgMottakToEsKlagebehandling(klagebehandlingOgMottak: Pair<Klagebehandling, Mottak>): EsKlagebehandling {
-
-        val klagebehandling = klagebehandlingOgMottak.first
-        //TODO: Nå bruker jeg ikke mottak her, så jeg kunne endret Pair<Klagebehandling, Mottak> til å bare være Klagebehandling?
-        //TODO: Er det noe vi skal indeksere opp som kommer fra Mottak? Beskrivelse f.eks?
-
-        val personInfo = klagebehandling.foedselsnummer?.let { pdlFacade.getPersonInfo(it) }
-        val erFortrolig = personInfo?.harBeskyttelsesbehovFortrolig() ?: false
-        val erStrengtFortrolig = personInfo?.harBeskyttelsesbehovStrengtFortrolig() ?: false
-        val erEgenAnsatt = klagebehandling.foedselsnummer?.let { egenAnsattService.erEgenAnsatt(it) } ?: false
-        val navn = personInfo?.navn ?: "mangler navn"
-
-        return EsKlagebehandling(
-            id = klagebehandling.id.toString(),
-            versjon = klagebehandling.versjon,
-            journalpostId = klagebehandling.saksdokumenter.map { it.referanse },
-            saksreferanse = klagebehandling.referanseId,
-            tildeltEnhet = klagebehandling.tildeltEnhet,
-            tema = klagebehandling.tema,
-            sakstype = klagebehandling.sakstype,
-            tildeltSaksbehandlerident = klagebehandling.tildeltSaksbehandlerident,
-            innsendt = klagebehandling.innsendt,
-            mottattFoersteinstans = klagebehandling.mottattFoersteinstans,
-            mottattKlageinstans = klagebehandling.mottattKlageinstans,
-            frist = klagebehandling.frist,
-            startet = klagebehandling.startet,
-            avsluttet = klagebehandling.avsluttet,
-            hjemler = klagebehandling.hjemler.map { it.original },
-            foedselsnummer = klagebehandling.foedselsnummer,
-            navn = navn,
-            egenAnsatt = erEgenAnsatt,
-            fortrolig = erFortrolig,
-            strengtFortrolig = erStrengtFortrolig
         )
     }
 }
