+++ src/main/kotlin/no/nav/klage/oppgave/clients/pdl/Person.kt
@@ -2,10 +2,10 @@ package no.nav.klage.oppgave.clients.pdl
 
 data class Person(
     val foedselsnr: String,
-    val fornavn: String,
+    val fornavn: String?,
     val mellomnavn: String?,
-    val etternavn: String,
-    val navn: String,
+    val etternavn: String?,
+    val navn: String?,
     val beskyttelsesbehov: Beskyttelsesbehov?
 ) {
     fun harBeskyttelsesbehovFortrolig() = beskyttelsesbehov == Beskyttelsesbehov.FORTROLIG
